refactor: add misc, core packages
