+++ src/misc/__init__.py
@@ -1,0 +1,2 @@
+from .constants import Color
+from .functions import from_tuple, generate_roman_number
